Break up users/api/api.go by feature
